Fix release build (#193)
PR #179 broke release builds by making it necessary for
`LiveAskQuestion` and `SEED` to both now be available regardless of
whether `debug_assertions` is set. That's unfortunate, since `SEED` in
particular would make the binary quite large. So I've instead made it so
that in release mode (or rather, without `debug_assertions`), `SEED` is
_not_ compiled in, and DynamoDB mode is _always_ used.

I've also added a CI job to check `--release --target
aarch64-unknown-linux-gnu` so that this can't happen in the future.